requirements.txt
Distutils2==1.0a4
M2Crypto  # Unversioned because it might be satisfied by the OS
Pygments==1.6
defusedxml==0.4.1
docutils==0.11
itsdangerous==0.23
passlib==1.6.1
psycopg2==2.5.1
py-bcrypt==0.4
python-dateutil==2.2
python-openid==2.2.5
pytz==2013.8
raven==3.5.2
redis==2.8.0
requests==2.0.1
rq==0.3.11
six==1.4.1
times==0.6.2
uWSGI==1.9.20
wsgiref==0.1.2
zope.browser==2.0.2
zope.component==4.1.0
zope.configuration==4.0.2
zope.contenttype==4.0.1
zope.event==4.0.2
zope.exceptions==4.0.6
zope.i18n==3.8.0
zope.i18nmessageid==4.0.2
zope.interface==4.0.5
zope.location==4.0.2
zope.pagetemplate==4.0.4
zope.proxy==4.1.3
zope.publisher==3.13.3
zope.schema==4.3.2
zope.security==4.0.0
zope.tal==3.6.1
zope.tales==4.0.2
zope.traversing==3.14.0
